+++ src/aligner.cpp
@@ -22,8 +22,8 @@ namespace xavier
 		hlength = hseq.length();
 		vlength = vseq.length();
 
-		queryh = new int8_t[hlength	+ VectorRegister::VECTORWIDTH];
-		queryv = new int8_t[vlength	+ VectorRegister::VECTORWIDTH];
+		queryh = new int16_t[hlength	+ VectorRegister::VECTORWIDTH];
+		queryv = new int16_t[vlength	+ VectorRegister::VECTORWIDTH];
 
 		std::copy(hseq.begin(), hseq.begin() + hlength, queryh);
 		std::copy(vseq.begin(), vseq.begin() + vlength, queryv);
@@ -88,7 +88,7 @@ namespace xavier
     	vqueryh[VectorRegister::LOGICALWIDTH] = VectorRegister::NINF;
     	vqueryv[VectorRegister::LOGICALWIDTH] = VectorRegister::NINF;
 
-        int antiDiagMax = std::numeric_limits<int8_t>::min();
+        int antiDiagMax = std::numeric_limits<int16_t>::min();
 
         // Load DPmatrix into antiDiag1 and antiDiag2 vector and
         // find max elem at the end of the initial stage in antiDiag1
@@ -149,9 +149,9 @@ namespace xavier
 			calcAntiDiag3();
 
 			// Update currScore
-			int8_t norm = updateCurrScore(); // currScore contains scoreOffset
+			int16_t norm = updateCurrScore(); // currScore contains scoreOffset
 
-			// Ensure anti-diagonals stay in int8_t range
+			// Ensure anti-diagonals stay in int16_t range
 	    	normalizeVectors(norm);
 
 	    	// Push back trace state
@@ -168,8 +168,10 @@ namespace xavier
 				return produceResults();
 			}
 
+			// std::cout << bestScore << std::endl;
+
 			// Update anti-diagonals
-			if (antiDiag3.argmax() > VectorRegister::LOGICALWIDTH / 2) moveRight();
+			if (antiDiag3.argmax() > VectorRegister::LOGICALWIDTH/2) moveRight();
 			else moveDown();
 		}
 
@@ -186,9 +188,9 @@ namespace xavier
 			calcAntiDiag3();
 
 			// Update currScore
-			int8_t norm = updateCurrScore();
+			int16_t norm = updateCurrScore();
 
-			// Ensure anti-diagonals stay in int8_t range
+			// Ensure anti-diagonals stay in int16_t range
 	    	normalizeVectors(norm);
 
 	    	// Trace state
@@ -265,9 +267,9 @@ namespace xavier
 		lastMove = DOWN;
 	}
 
-	int8_t Aligner::updateCurrScore()
+	int16_t Aligner::updateCurrScore()
 	{
-		int8_t antiDiagBest = *std::max_element( antiDiag3.internal.elems,
+		int16_t antiDiagBest = *std::max_element( antiDiag3.internal.elems,
 		                                         antiDiag3.internal.elems
 		                                          + VectorRegister::VECTORWIDTH );
 		currScore = antiDiagBest + scoreOffset;
@@ -286,7 +288,7 @@ namespace xavier
 		return hoffset > hlength || voffset > vlength;
 	}
 
-	void Aligner::normalizeVectors(int8_t& normfactor)
+	void Aligner::normalizeVectors(int16_t& normfactor)
 	{
 		int64_t antiDiagBest = currScore - scoreOffset;
